Add psp build .pbp

+++ src/SF_EXE/resources/retail_filesystem.cpp
@@ -87,7 +87,11 @@ std::int32_t countRetailSaves(
     std::int32_t count = 0;
     for (std::int32_t index = 0; index < 6; ++index) {
         char path[32]{};
-        std::snprintf(path, sizeof(path), "Save\\%04d.Ssv", index);
+        std::snprintf(
+            path,
+            sizeof(path),
+            "Save\\%04d.Ssv",
+            static_cast<int>(index));
         if (retailFileExists(root, path)) {
             ++count;
         }
@@ -104,7 +108,10 @@ bool deleteRetailSave(
     for (std::int32_t slot = 0; slot < 6; ++slot) {
         char save_path[32]{};
         std::snprintf(
-            save_path, sizeof(save_path), "Save\\%04d.Ssv", slot);
+            save_path,
+            sizeof(save_path),
+            "Save\\%04d.Ssv",
+            static_cast<int>(slot));
         if (!retailFileExists(root, save_path)) {
             continue;
         }
@@ -120,7 +127,7 @@ bool deleteRetailSave(
             preview_path,
             sizeof(preview_path),
             "Save\\%04d.Bmp",
-            slot);
+            static_cast<int>(slot));
         error.clear();
         std::filesystem::remove(
             resolveRetailPath(root, preview_path), error);
